backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
twilio==9.0.0
openai==1.52.2
python-dateutil==2.9.0.post0
pytz==2024.2
celery==5.4.0
redis==5.2.0
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0